[m01/ex04] Use indices and length instead of iterators
While iterators are more "C++ way", they make the code harder to
understand and should probably be considered a premature generalization
in this case - especially as the `to_match` and `replacement` arguments
are still strings and not pairs of begin and end iterators.

Using indices and length allows us to call `ostream::write` directly
instead of calling `ostream::put` in a loop.