[m03/ex03] Use static constants for initializing attributes in ScavTrap and FragTrap
This allows a class inheriting both ScavTrap and FragTrap to access
constants from each and choose which values to use without hardcoding
them. Relying on ScavTrap and FragTrap constructors would not allow
choosing a mix as the constructor that runs second overwrites the values
in the shared ClapTrap base object.