[m04/ex03] Avoid reassigning AMateria::type
As hinted in the subject, this assignment doesn't make sense as
assignment should only be allowed between matching types, and in this
case the AMateria::type variable should have identical values.

If AMateria::type field is not reassigned, there is no reason to call or
even allow calling AMateria::operator= at all.